fix: Calculating touch positions according to scaling of canvas.
