+++ src/main.ts
@@ -23,56 +23,17 @@
  */
 
 import "./style.css";
-import { canvas } from "./graphics";
 import { init } from "./game";
+import { resizeCanvasMaintainingAspectRatio } from "./window";
 
 const maxWidth = 1280;
 const maxHeight = 720;
 
-const resize = (): void => {
-    // Calculate the aspect ratio
-    const aspectRatio = maxWidth / maxHeight;
-
-    // Calculate the width and height based on the window size while maintaining the aspect ratio
-    let width = window.innerWidth;
-    let height = window.innerHeight;
-
-    if (width / height > aspectRatio) {
-        width = height * aspectRatio;
-    } else {
-        height = width / aspectRatio;
-    }
-
-    // Ensure the width and height do not exceed the maximum resolution
-    if (width > maxWidth) {
-        width = maxWidth;
-        height = maxWidth / aspectRatio;
-    }
-    if (height > maxHeight) {
-        height = maxHeight;
-        width = maxHeight * aspectRatio;
-    }
-
-    // Set the canvas width and height
-    canvas.width = width;
-    canvas.height = height;
-
-    // Calculate the scale factor to fill the screen
-    const scaleX = window.innerWidth / width;
-    const scaleY = window.innerHeight / height;
-    const scale = Math.min(scaleX, scaleY);
-
-    // Center the canvas
-    canvas.style.position = "absolute";
-    canvas.style.left = `${(window.innerWidth - width * scale) / 2}px`;
-    canvas.style.top = `${(window.innerHeight - height * scale) / 2}px`;
-
-    // Apply the scaling
-    canvas.style.transform = `scale(${scale})`;
-    canvas.style.transformOrigin = "top left";
-};
-
-window.addEventListener("resize", resize, false);
-resize();
+window.addEventListener(
+    "resize",
+    () => resizeCanvasMaintainingAspectRatio(maxWidth, maxHeight),
+    false,
+);
+resizeCanvasMaintainingAspectRatio(maxWidth, maxHeight);
 
 init();
